Refactors Requests interfaces
