--- src/models/responses/Brand.ts
@@ -1,5 +1,0 @@
-import Model from "./Model";
-
-export default interface Brand extends Model {
-
-}
