--- src/models/responses/Category.ts
@@ -1,5 +1,0 @@
-import Model from "./Model";
-
-export default interface Category extends Model {
-
-}
