--- src/models/responses/ItemAnalyticsDTO.ts
@@ -1,9 +1,0 @@
-import Item  from "./Item";
-
-export default interface ItemAnalyticsDTO {
-    item: Item,
-    latestStore: String,
-    latestDate: String,
-    latestPrice: String,
-    averagePrice: String
-}
