--- src/models/responses/Model.ts
@@ -1,4 +1,0 @@
-export default interface Model {
-    id: number,
-    name: string
-}
