--- src/models/responses/Response.ts
@@ -1,3 +1,0 @@
-export interface Response<T>{
-    data: T
-}
