--- src/models/responses/Store.ts
@@ -1,5 +1,0 @@
-import Model from "./Model";
-
-export default interface Store extends Model {
-
-}
